refactor(gsd-db): peel memory writers into db/writers/memory.ts
Phase A, commit 2. Extracts the 14 memory-store write wrappers (the file's
self-contained tail) into db/writers/memory.ts, re-exported by the barrel.

Re-scopes the single-writer invariant from a basename allowlist to a directory
predicate: write SQL is permitted under db/engine.ts and db/writers/**, plus
gsd-db.ts (barrel) and unit-ownership.ts. This is the candidate-2 enforcement
change — the invariant now guards a seam, not a filename.

Verified: typecheck:extensions clean; single-writer-invariant + memory
sources/ingest/store round-trip tests green (56 tests).
